add drag and drop on certain position
when user drags the email inside a collider it gets isInisde value to true and takes the tag of the collider (either Phishing or Legitimate), if email exits the collider while the user is still holding it isInside is set to false. The script is applied to Email Object where isPhishing is set to true/false according to whether the email is phishing. There is a check if isPhishing is true and if the email object was dropped inside Legitimate/Phishing collider and answer is provided according to this. Currently there are only logs, icons will be added on true/false